requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
asyncpg==0.31.0
attrs==25.4.0
audioop-lts==0.2.2
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
comtypes==1.4.15
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
edge-tts==7.2.7
elevenlabs==2.33.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.109.0
filelock==3.20.3
Flask==3.0.0
frozenlist==1.8.0
fsspec==2026.1.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
llvmlite==0.46.0
Mako==1.3.10
MarkupSafe==3.0.3
more-itertools==10.8.0
motor==3.7.1
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
ollama==0.6.1
openai==1.12.0
openai-whisper==20250625
passlib==1.7.4
propcache==0.4.1
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
PyJWT==2.11.0
pymongo==4.16.0
pypiwin32==223
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.20
pyttsx3==2.99
pywin32==311
regex==2026.1.15
requests==2.32.5
rsa==4.9.1
scipy==1.17.0
setuptools==80.10.2
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
soundfile==0.13.1
SQLAlchemy==2.0.46
starlette==0.35.1
structlog==25.5.0
sympy==1.14.0
tabulate==0.9.0
tiktoken==0.12.0
tomli==2.4.0
torch==2.10.0
tqdm==4.67.2
twilio==9.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.27.0
websockets==12.0
Werkzeug==3.1.5
whisper==1.1.10
yarl==1.22.0
